Block API: refactor.
