+++ Source/Ivxr.SePlugin/IvxrPluginContext.cs
@@ -29,7 +29,8 @@ namespace Iv4xr.SePlugin
 			var sessionDispatcher = new SessionDispatcher(sessionController) { Log = Log };
 
 			m_server = new PluginServer(Log, sessionDispatcher, m_requestQueue);
-			var observer = new Observer(m_gameSession) { Log = Log };
+			var llObserver = new LowLevelObserver(m_gameSession) { Log = Log };
+			var observer = new Observer(llObserver) { Log = Log };
 			var controller = new CharacterController(m_gameSession);
 
 			Dispatcher = new Dispatcher(m_requestQueue, observer, controller) { Log = Log };
